Improve performance of prime generator
The previous implementation was an exercise in recursion and I thought
it was an interesting novelty to have all three prime related functions
(is_prime, primes and prime_factors) interdependent in some way.

This was, however, slowing down tests, so after some research I've
reimplemented the primes generator as a sieve to get the tests to run faster.