[problem_0015] Initial commit
Commiting this just for posterity because the solution is brute force and
quickly binds up. I couldn't get an answer for the question "how many paths
through a 20 * 20 lattice" with this method, but I liked the simple logic
of the code, so it seemed a shame not to keep it somewhere.